ImageServer/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cors==1.0.1
cycler==0.11.0
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
gdown==4.7.1
gevent==22.10.2
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opennsfw2==0.10.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tifffile==2023.3.21
tldextract==3.4.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
